requirements.txt
atlassian-python-api==3.41.13
beautifulsoup4==4.12.3
jira==3.8.0
g4f==0.3.1.0
lxml==5.2.2
psycopg2-binary==2.9.9
requests==2.31.0
telebot==0.0.5
schedule==1.2.2
PyYAML==6.0.2
testit-api-client==4.0.0